refactor: split Component tests
